SOL/token amount - Data Type
- Changed token_amount fields from f64 to u64 in StrategyPosition, Trade, and related models to ensure exact integer representation.
- Updated methods for setting and retrieving token amounts to accommodate the new u64 type.
- Adjusted conversion functions for SOL to lamports and vice versa, ensuring precision is maintained.
- Modified database repository bindings to reflect changes in data types, ensuring compatibility with the new schema.
- Updated tests to reflect changes in token amount handling, ensuring correctness in calculations and assertions.

+++ ingest-laserstream/src/decode/create.rs
@@ -65,7 +65,7 @@ impl Decoder {
             .iter()
             .find(|ev| ev.is_buy && ev.user == buy_user && ev.mint == mint);
 
-        let initial_supply = initial_create_event.map(|ev| ev.token_amount as u64);
+        let initial_supply = initial_create_event.map(|ev| ev.token_amount);
         let initial_buy_sol = initial_create_event.map(|ev| ev.sol_amount);
 
         let name = create_log.map(|e| e.name.clone())
