+++ trading_core/src/models/ingest.rs
@@ -50,7 +50,9 @@ pub enum SseEvent {
         wallet: String,
         trade_type: TradeType,
         sol_amount: f64,
-        token_amount: f64,
+        /// Raw token units — exact `u64` (serializes as a JSON number; the frontend
+        /// scales to a display amount). No `f64` round-trip on the notify path.
+        token_amount: u64,
         price_per_token: f64,
         tx_signature: String,
         slot: u64,
